// dashboard/src/utils/hasura-api/generated/schemas/customRootField.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */

export interface CustomRootField {
  /**
   * The custom root field name
   * @nullable
   */
  name?: string | null;
  /** Customize the description shown for the root field in GraphQL introspection. If null or omitted then an autogenerated description is used instead. */
  comment?: string;
}